DOH
My tired little head literally told you in the last commit why it was
happening and was too dumb to realise it - the scroller is appearing
because there's enough content on those pages to require one. Once the
site is filled out a bit, this'll be a distant memory!

Also, I've added little thumbnails to the social media dropdown links