x4iiiis.com Back From The Dead
I *THINK* I've brought it back to how it was. Somehow fucked up and deleted pretty much everything cause it was saved on icloud rather than on disk, and when I went to Manchester I think I fucked it. Github wasn't up to date, so it was a bit of a fucking nightmare. Also, in FileZilla, public_html no longer exists, so I may have fucked up. I dunno. :confused: